refactor: address PR #190 review — compose package, gate base fix, doc trims
Review follow-ups on the Project() redesign:

- compose.py → compose/ package: errors.py (ProjectLoadError), rebase.py
  (cwd/paths/when re-anchoring), scope.py (the isolated _compose_scope +
  loaders). Non-recursive main/*.py glob leaves the package interpreted, so
  compose.py drops out of setup.py _main_excluded.
- rename load_py_state → load_py_tasks_state.
- gate_cli: derive the rebasing frame from the resolved tasks.py dir (base_from,
  now shared with base_for) and thread base into run_gate, matching the MCP
  server path — a leaf spawned from a child cwd now anchors to its tasks.py.
- execution.py: Path import → TYPE_CHECKING (drop noqa, doctest imports it);
  document why RunContext.base can be None.
- drop the compose module-docstring slop; fix run_cli's stale "discovered
  descendants" wording (discovery is gone).
- tests: gate cwd-from-subdir anchoring; inline Project composes but mounts no
  namespace (no binding name to mount under).
